fix zsh/main.sh: make Powerline install idempotent and use subshell for cd
- Skip cloning and installing Powerline fonts if they're already present
- Use a subshell (cd fonts && ./install.sh) instead of cd/cd.. so a
  failure inside install.sh can't leave the script in the wrong directory

+++ zsh/main.sh
@@ -34,11 +34,8 @@ log "Setting custom Zsh theme"
 cp -rf "./zsh/jsegal_theme.zsh-theme" "$ZSH/custom/themes/jsegal_theme.zsh-theme"
 
 log "Installing Powerline Fonts"
-# clone
-git clone https://github.com/powerline/fonts.git --depth=1
-# install
-cd fonts
-./install.sh
-# clean-up a bit
-cd ..
-rm -rf fonts
+if [ ! -d "$HOME/.local/share/fonts" ] || ! fc-list | grep -q "Powerline"; then
+  git clone https://github.com/powerline/fonts.git --depth=1
+  (cd fonts && ./install.sh)
+  rm -rf fonts
+fi
